Remove dependency on Swift-NIO SSL
This removes Swift-DocC's dependency on Swift-NIO SSL.

The removal is motivated by Swift-NIO SSL deprecating support for CentOS 7 due
to BoringSSL removing support for CentOS 7.

The Swift-NIO SSL dependency is currently used to support running a
documentation preview server over HTTPS which allows users to preview content
on a secondary device while authoring it on a primary device since
many OS's won't allow accessing non-local content over HTTP.

With the introduction of `--transform-for-static-hosting` and the subsequent
enabling of that feature by default, there's less need for Swift-DocC to ship
a solution for this niche feature out-of-the-box since standard command-line
based preview servers now support serving DocC archives with minimal additional
configuration.

Resolves rdar://93197483.